testing: create testing folder and move test utils
