+++ src/app/components/ui/button/button.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import {
   getElementBySelector,
   getElementTextContentBySelector,
-} from '@app/utils';
+} from '@app/testing';
 import { TranslateModule } from '@ngx-translate/core';
 
 import { ButtonComponent } from './button.component';
