+++ src/app/components/ui/spinner/spinner.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { getElementBySelector } from '@app/utils';
+import { getElementBySelector } from '@app/testing';
 import { TranslateModule } from '@ngx-translate/core';
 
 import { SpinnerComponent } from './spinner.component';
