+++ src/app/components/user/list/user-list.component.spec.ts
@@ -18,7 +18,7 @@ import {
   getAllElementsBySelector,
   getElementBySelector,
   getElementTextContentBySelector,
-} from '@app/utils';
+} from '@app/testing';
 import { Store } from '@ngrx/store';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { TranslateModule } from '@ngx-translate/core';
