+++ src/app/components/user/post/user-post.component.spec.ts
@@ -14,7 +14,7 @@ import {
   getAllElementsTextContentBySelector,
   getElementBySelector,
   getElementTextContentBySelector,
-} from '@app/utils';
+} from '@app/testing';
 import { Store } from '@ngrx/store';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { TranslateModule } from '@ngx-translate/core';
