+++ src/app/components/user/profile/user-profile.component.spec.ts
@@ -16,7 +16,7 @@ import {
   getAllElementsTextContentBySelector,
   getElementBySelector,
   getElementTextContentBySelector,
-} from '@app/utils';
+} from '@app/testing';
 import { Store } from '@ngrx/store';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { TranslateModule } from '@ngx-translate/core';
